wallet, topup, payment method

+++ database/migrations/2024_01_16_135047_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('category_id')->constrained();
             $table->string('name');
-            $table->float('price');
+            $table->unsignedBigInteger('price');
             $table->longText('description');
             $table->string('tags')->nullable();
             $table->timestamps();
